Split auth to auth.rs
